backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
openai==1.50.0
edge-tts==6.1.12
moviepy==1.0.3
aiofiles==24.1.0
python-dotenv==1.0.1
yt-dlp==2024.8.6
Pillow==10.4.0
